[18-ndcg-smallfloat] feat(scoring): quantize doc_len via Lucene SmallFloat (parite NDCG TREC-COVID + ~65 MiB)
Ferme la parite stricte qualite : TREC-COVID NDCG@10 0.4750 vs OS 0.4902
(-0.0152) etait entierement explique par le scoring BM25 cote Surch qui
utilisait des doc_len exacts (Vec<u64>) la ou Lucene utilise une
quantization 1 byte/doc via SmallFloat.intToByte4 / byte4ToInt. A taille
de top-K identique (Recall@10 identique), l ordre interne divergeait sur
les paires que Lucene regroupe dans le meme bucket. Voir
docs/paper/ndcg-trec-covid-rootcause-22.md (#22).

- crates/surch-search/src/small_float.rs : port bit-pour-bit de
  org.apache.lucene.util.SmallFloat (4-bit mantissa + 4-bit exponent +
  free range [0..24)). Reference canonique + tests sur valeurs critiques
  hand-traces depuis la source Lucene (0, 1, 24, 25, 31, 32, 255, 1000,
  1500, 65535) + invariants (free range identite, decoder monotone,
  bucket-collapse, BM25 parite vs reference manuelle).
- crates/surch-index/src/document_index.rs : FieldLengthStats.doc_len_dense
  passe de Vec<u64> -> Vec<u8>. Mirror prive du codec SmallFloat (le crate
  surch-search depend deja de surch-index, import inverse impossible).
  record_doc_len quantize a l ecriture, doc_len(doc_id) et min_doc_len()
  retournent la longueur reconstruite (meme valeur que Lucene). total_terms
  reste la somme RAW (avg_doc_len matche sumTotalTermFreq / docCount cote
  Lucene). Expose decode_doc_len_byte() pour les consommateurs zero-copy.
- crates/surch-search/tests/small_float_parity.rs : gate CI byte-identite
  entre les deux implementations sur la plage des longueurs reelles
  (1..=4096 + boundary cases), via round-trip DocumentIndex.
- crates/surch-api/src/state.rs : FieldScoringStats.doc_len_dense bascule
  &[u64] -> &[u8] (toujours zero-copy). doc_len(doc_id) decode via
  surch_index::decode_doc_len_byte. min_doc_len reste deja dans le domaine
  reconstruit cote surch-index donc inchange cote API.
- crates/surch-index/src/memory.rs : entry_size ledger u64 -> u8 (8x
  reduction de la ligne field_stats_bytes).

Gains memoire attendus :
- deces (1.36M docs x ~6 champs analyses) : ~65 MiB economises sur
  field_stats_bytes (qui pesait 126 MiB, soit ~52 % de reduction).
- TREC-COVID (171k docs) : ~8 MiB.

Garde-fous :
- Tests existants document_index.rs doc_len(0) == Some(3) etc. restent
  verts car la plage [0, 24) est passthrough lossless (raw < 24 -> raw).
- Parite oracle b1/b2 deces : la quantization repose sur un encoder
  deterministe. Les paires actuellement departagees par doc_len exact qui
  tombent dans le meme bucket Lucene sont alors departagees par le tie-
  break par doc_id (deja en place), donc 0-divergence preservee.
- SciFact (corpus court, lengths < 200) majoritairement dans la plage
  passthrough -> ndcg-gate >= 0.6537 doit rester verte (actuellement
  0.6576, marge OK).

Validation : ce changement N A PAS ete teste en local (cargo test
plante le PC user, 5 fois deja). Validation = CI cluster apres push.

+++ crates/surch-index/src/memory.rs
@@ -219,7 +219,11 @@ fn subfield_values_bytes(doc_index: &DocumentIndex) -> u64 {
 
 fn field_stats_bytes(doc_index: &DocumentIndex) -> u64 {
     let stats_header = size_of::<FieldLengthStats>() as u64;
-    let entry_size = size_of::<u64>() as u64;
+    // #18 ndcg-smallfloat: `doc_len_dense` switched from `Vec<u64>` (8 B/doc)
+    // to `Vec<u8>` (Lucene `SmallFloat`-quantized, 1 B/doc). The ledger
+    // entry size shrinks 8× — ~65 MiB freed on the deces 1.36 M × ~6
+    // indexed fields workload.
+    let entry_size = size_of::<u8>() as u64;
     let mut total: u64 = 0;
     for (field, stats) in doc_index.field_stats_map().iter() {
         total += field.len() as u64;
